@@ -4,21 +4,21 @@ android {
     signingConfigs {
         release {
             storeFile file('E:/ak95/My/androidsign')
-            storePassword 'lloret95'
+            storePassword ''
             keyAlias "sign"
-            keyPassword "lloret95"
+            keyPassword ""
         }
     }
 
-    compileSdkVersion 27
-    buildToolsVersion "27.0.3"
+    compileSdkVersion 29
+    buildToolsVersion '29.0.2'
 
     defaultConfig {
         applicationId "com.antonklimakov.metar"
         minSdkVersion 16
-        targetSdkVersion 27
-        versionCode 7
-        versionName "1.1"
+        targetSdkVersion 29
+        versionCode 8
+        versionName "1.2"
         resConfigs "en", "ru"
     }
     buildTypes {
@@ -37,8 +37,7 @@ android {
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'com.android.support:support-v4:27.0.2'
-    implementation 'com.google.code.gson:gson:2.8.2'
+    implementation 'com.google.code.gson:gson:2.8.5'
     implementation 'org.jsoup:jsoup:1.11.2'
-    implementation 'com.github.apl-devs:appintro:v4.2.3'
+    implementation 'com.github.AppIntro:AppIntro:5.1.0'
 }
