@@ -1,9 +1,9 @@
 package com.antonklimakov.metar;
 
-import android.graphics.Color;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
+
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 
 import com.antonklimakov.metar.fragments.SampleSlide;
 import com.github.paolorotolo.appintro.AppIntro;
